refactor: Move navbar, sidebar, and notification button components to layoutComponents
